style: move directory
