+++ backend/src/test/java/com/hjjang/backend/domain/user/service/UserProfileServiceTest.java
@@ -5,7 +5,7 @@ import com.hjjang.backend.domain.user.entity.Agreement;
 import com.hjjang.backend.domain.user.entity.RoleType;
 import com.hjjang.backend.domain.user.entity.User;
 import com.hjjang.backend.domain.user.repository.UserRepository;
-import com.hjjang.backend.security.WithMockCustomUser;
+import com.hjjang.backend.global.security.WithMockCustomUser;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
