+++ backend/src/test/java/com/hjjang/backend/global/security/WithMockUserSecurityContextFactory.java
@@ -1,4 +1,4 @@
-package com.hjjang.backend.security;
+package com.hjjang.backend.global.security;
 
 import com.hjjang.backend.domain.user.entity.RoleType;
 import com.hjjang.backend.global.config.security.properties.AuthProperties;
